// src/app/shared/models/i-player-registration.ts
export interface IPlayerRegistration {
  name: string;
  mobile: number;
  motherName: string;
  fatherName: string;
  dob: string;
  education: string;
  school: string;
  district: string;
  vikaskhand: string;
  nyayPanchyat: string;
  gramPanchyat: string;
  aadharNumber: number;
  photoName: string;
  aadharPhoto?: string;
}